requirements.txt
mysqlclient==1.4.4
PyQt5==5.13.1
PyQt5-sip==4.19.19
